Implement the env

+++ _getenv.c
@@ -1,4 +1,4 @@
-#include "main.h"
+#include "shell.h"
 
 /**
  * _getenv - function that returns environment value
@@ -12,13 +12,13 @@ char *_getenv(char *variable)
 {
 	size_t i, j;
 	char *envstr;
-	int len = strlen(variable);
+	size_t len = strlen(variable);
 
 	for (i = 0; environ[i]; i++)
 	{
 		envstr = environ[i];
 		/*compare variable with envstr*/
-		for(j = 0; j < len; j++)
+		for (j = 0; j < len; j++)
 		{
 			if (envstr[j] == variable[j])
 				continue;
